requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
fastapi==0.116.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.5.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langchainhub==0.1.21
langsmith==0.3.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.28.0
networkx==3.5
numpy==1.26.4
openai==1.64.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.2.0
postgrest==0.19.3
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.47.1
storage3==0.11.3
streamlit==1.42.2
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.53.2
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==14.2
yarl==1.18.3
zstandard==0.23.0
